+++ neural_lam/models/latent/graph_decoder.py
@@ -30,6 +30,8 @@ class GraphLatentDecoder(BaseGraphLatentDecoder):
         g2m_gnn_type="InteractionNet",
         m2g_gnn_type="InteractionNet",
         output_std=True,
+        num_grid_nodes=None,
+        num_mesh_nodes=None,
     ):
         """
         Set up the g2m, on-mesh and m2g GNNs.
@@ -72,13 +74,18 @@ class GraphLatentDecoder(BaseGraphLatentDecoder):
             hidden_dim,
             hidden_layers=hidden_layers,
             update_edges=False,
+            num_rec=num_mesh_nodes,
         )
 
         # None if m2m_layers == 0, in which case no on-mesh processing is
         # done in combine_with_latent
         self.m2m_gnns = (
             utils.make_gnn_seq(
-                m2m_edge_index, m2m_layers, hidden_layers, hidden_dim
+                m2m_edge_index,
+                m2m_layers,
+                hidden_layers,
+                hidden_dim,
+                num_rec=num_mesh_nodes,
             )
             if m2m_layers > 0
             else None
@@ -89,6 +96,7 @@ class GraphLatentDecoder(BaseGraphLatentDecoder):
             hidden_dim,
             hidden_layers=hidden_layers,
             update_edges=False,
+            num_rec=num_grid_nodes,
         )
 
     def combine_with_latent(